+++ clown_sort/config.py
@@ -40,6 +40,7 @@ class Config:
     leave_in_place: bool = False
     screenshots_only: bool = True
     delete_originals: bool = False
+    rescan_sorted: bool = False
     sort_rules: List[SortRule] = []
     filename_regex: re.Pattern
 
@@ -66,6 +67,7 @@ class Config:
         Config.filename_regex = re.compile(args.filename_regex)
         Config.leave_in_place = True if args.leave_in_place else False
         Config.delete_originals = True if args.delete_originals else False
+        Config.rescan_sorted = True if args.rescan_sorted else False
 
         if Config.leave_in_place and Config.delete_originals:
             Console().print("--leave-in-place and --delete-originals are mutually exclusive.", style='red')