+++ clown_sort/util/argument_parser.py
@@ -51,6 +51,9 @@ parser.add_argument('-f', '--filename-regex',
                     help='filename regular expression',
                     default=DEFAULT_FILENAME_REGEX.pattern)
 
+parser.add_argument('--rescan-sorted', action='store_true',
+                    help="rescan already sorted files (developer option)")
+
 parser.add_argument('--delete-originals', action='store_true',
                     help="don't preserve the original screenshots in the Processed/ folder")